+++ run_celery_beat.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
-rm -f /usr/src/app/*.pid
+rm -f /code/*.pid
 sleep 10
 celery -A icloud.celery beat -l debug